Fix Electron runtime: CommonJS output, renderer sandboxing, update AGENTS.md
- Fix tsconfig.json: use commonjs module output (not ES2022) for Electron compat
- Fix renderer.ts: move type declarations to global.d.ts to avoid CommonJS
  'exports' reference in sandboxed browser context
- Fix renderer.ts: rename 'api' variable to 'songApi' to avoid collision
  with contextBridge-exposed window.api (read-only property)
- Fix dev-server.ts: use __dirname instead of import.meta.url (CommonJS)
- Fix index.html: remove type="module" from script tag
- Update AGENTS.md with working Electron runtime instructions and gotchas
