+++ electron-src/renderer/global.d.ts
@@ -1,0 +1,13 @@
+interface JWorshipAPI {
+  listSongs(): Promise<string[]>;
+  loadSong(fileName: string): Promise<{
+    fileName: string;
+    title: string;
+    author: string;
+    verses: string[];
+  } | null>;
+}
+
+interface Window {
+  api: JWorshipAPI;
+}
