+++ electron-src/renderer/index.html
@@ -30,6 +30,6 @@
       </section>
     </main>
   </div>
-  <script type="module" src="renderer.js"></script>
+  <script src="renderer.js"></script>
 </body>
 </html>
